Rename NewPackageGraphBuilder with a code clean up
